+++ test/error_test.py
@@ -34,6 +34,11 @@ samila.errors.samilaDataError: Data file can't be saved. At least one of the dat
 Traceback (most recent call last):
         ...
 samila.errors.samilaPlotError: Plotting process can't be Done because data2 is empty. Use generate method first.
+>>> g.generate()
+>>> g.plot(color=(1, 2, 3, 4, 5))
+Traceback (most recent call last):
+        ...
+samila.errors.samilaPlotError: Given color type is not supported.
 >>> g = GenerativeImage(lambda x,y: x, lambda x,y: y)
 >>> result = g.save_config()
 Traceback (most recent call last):